# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/records": {
            "post": {
                "description": "Create a new record associated with a specific reading",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReadingRecord"
                ],
                "summary": "Add a new record for a reading",
                "parameters": [
                    {
                        "description": "Reading Record details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ReadingRecordInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Reading Record created",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Reading not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/records/{readingId}": {
            "get": {
                "description": "Retrieve all records from a user's reading by reading ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReadingRecord"
                ],
                "summary": "Get records for a specific reading",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reading ID",
                        "name": "readingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found reading records",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of a specific record for a reading",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReadingRecord"
                ],
                "summary": "Update an existing record for a reading",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reading ID",
                        "name": "readingId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Reading Record details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ReadingRecordInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reading Record updated",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Reading not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        },
        "/records/{recordId}": {
            "put": {
                "description": "Update the details of a reading record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReadingRecord"
                ],
                "summary": "Update an existing reading record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Reading Record details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ReadingRecordInsertDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reading Record updated",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Reading not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific reading record from a user's readings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReadingRecord"
                ],
                "summary": "Remove a record from my readings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reading ID",
                        "name": "readingId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reading Record deleted",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Reading not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.ReadingRecordInsertDTO": {
            "type": "object",
            "required": [
                "notes",
                "progress",
                "reading_id"
            ],
            "properties": {
                "notes": {
                    "type": "string"
                },
                "progress": {
                    "type": "string"
                },
                "reading_id": {
                    "type": "string"
                }
            }
        },
...
        "utils.ApiResponse": {
            "description": "Represents a standardized API response structure.",
            "type": "object",
            "properties": {
                "code": {
                    "description": "The HTTP status code",
                    "type": "integer"
                },
                "data": {
                    "description": "The data returned by the API"
                },
                "msg": {
                    "description": "A message providing additional context",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates if the request was successful",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "The time at which the response was generated",
                    "type": "string"
                }
            }
        }
    }
}